Add structure to Compose components
